COM-1370 add course order and status display
